delete post Handler bug fixed

+++ controllers/feed.js
@@ -1,11 +1,12 @@
 const fs = require('fs');
 const path = require('path');
 //importing validation result from the package
-//const {validationResult} = require('express-validator/check');
-const {validationResult} = require('express-validator');
+const {validationResult} = require('express-validator/check');
+//const {validationResult} = require('express-validator');
 
 //importing post from models
 const Post = require('../models/post');
+const User = require('../models/user');
 
 //exporting function (will return a json response)
 exports.getPosts = (req, res, next) => {
@@ -32,7 +33,7 @@ exports.getPosts = (req, res, next) => {
     });
 };
 
-//post router
+//post router to create blog post
 exports.createPost = (req, res, next) => {
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
@@ -48,23 +49,31 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path;
     const title = req.body.title;
     const content = req.body.content;
-
+    let creator;
     //post model constructor 
     const post = new Post({
         title: title, 
         content: content, 
         imageUrl: imageUrl,
-        creator:{ name: 'Deisiane' },
+        creator: req.userId, //assign post to the user 
     });
     post
         .save()
         .then(result => {
-            console.log(result);
-        //create post in db 
-        //setting status method to 201 (tell the client  the resource was created success) / 200 (sucesso only)
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        //console.log(result)
+        })
+        .then(result => {
+            //setting status method to 201 (tell the client  the resource was created success) / 200 (sucesso only)
             res.status(201).json({
                 message: 'Your Post was created successfully!',
-                post:  result
+                post: post,
+                creator: {_id: creator._id, name: creator.name}
             });
         })
         .catch(err => {
@@ -123,6 +132,11 @@ exports.updatePost = (req, res, next) => {
             error.statusCode = 404;
             throw error;
         }
+        if (post.creator.toString() !== req.userId) {
+            const error = new Error ('Not Authorized!');
+            error.statusCode = 403;
+            throw error;
+        }
         if (imageUrl !== post.imageUrl) {
             clearImage(post.imageUrl);
         }
@@ -152,13 +166,24 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
             //check logged in user
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
         .then(result => {
-            console.log(result);
-            res.status(200).json({ message: 'Post was deleted successfully!'})
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then(result => {
+            res.status(200).json({ message: 'Post was deleted successfully!'});
         })
         .catch(err => {
             if (!err.statusCode) {
